Add Hero-section and organize it w/ flex
After adding all the basic html elements for the hero section,
classes were added. First main class was a container for the hero
section. This class was given the display of 'flex'. Within the
container, another div with the class hero-section-text was added.
Within that class was the main hero title, paragraph, and button.
After that div class was the image which was given a width attribute
of 400px.